+++ src/components/onboarding/index.ts
@@ -1,0 +1,3 @@
+export { OnboardingProvider, useOnboarding, useOnboardingOptional } from "./OnboardingProvider";
+export { default as OnboardingTour } from "./OnboardingTour";
+export { default as OnboardingTrigger } from "./OnboardingTrigger";